asymptote: remove display side-effect
When building the documentation, the call to asymptote outputs the pwd
and the checksum; this interrupts the new progress bar mechanism in
recent Racket versions. Silence it by removing the display entirely.

(Perhaps we should provide a function to get the hash or document how it
is computed? The function could also accept a side-channel to
communicate the hash to, such as a box. Modifying it to return multiple
values would be a breaking change.)